[Core] add `xgrammar` as guided generation provider
Essentially a cleaned up version of this `pr`:
https://github.com/vllm-project/vllm/pull/10785

Especially since `outlines` is rather slow and the new version is though
to intergrate as they do not focus on being pickleable which is a key
feature for us using the multiprocessing engine: https://github.com/dottxt-ai/outlines-core/pull/99

I assume more and more will change over to `xgrammar`.

This is a minimum implementation.

https://arxiv.org/pdf/2411.15100

Signed-off-by: Jannis Schönleber <joennlae@gmail.com>